SYNC-1571 Fix split package
