+++ portlets/sync-admin-portlet/docroot/WEB-INF/src/com/liferay/sync/admin/portlet/AdminPortlet.java
@@ -28,7 +28,7 @@ import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PrefsPropsUtil;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.sync.OAuthPortletUndeployedException;
+import com.liferay.sync.admin.OAuthPortletUndeployedException;
 import com.liferay.sync.service.SyncPreferencesLocalServiceUtil;
 import com.liferay.sync.shared.util.PortletPropsKeys;
