+++ portlets/sync-admin-portlet/docroot/init.jsp
@@ -46,7 +46,7 @@ page import="com.liferay.portal.kernel.util.StringPool" %><%@
 page import="com.liferay.portal.kernel.util.StringUtil" %><%@
 page import="com.liferay.portal.kernel.util.Validator" %><%@
 page import="com.liferay.portal.kernel.util.WebKeys" %><%@
-page import="com.liferay.sync.OAuthPortletUndeployedException" %><%@
+page import="com.liferay.sync.admin.OAuthPortletUndeployedException" %><%@
 page import="com.liferay.sync.service.SyncPreferencesLocalServiceUtil" %><%@
 page import="com.liferay.sync.shared.util.PortletPropsKeys" %><%@
 page import="com.liferay.sync.shared.util.SyncPermissionsConstants" %>
